+++ sdk/storage/azure-storage-blobs/src/append_blob_client.cpp
@@ -111,8 +111,10 @@ namespace Azure { namespace Storage { namespace Blobs {
           && e.ErrorCode == "BlobAlreadyExists")
       {
         Models::CreateAppendBlobResult ret;
+        ret.RequestId = e.RequestId;
         ret.Created = false;
-        return Azure::Core::Response<Models::CreateAppendBlobResult>(std::move(ret), std::move(e.RawResponse));
+        return Azure::Core::Response<Models::CreateAppendBlobResult>(
+            std::move(ret), std::move(e.RawResponse));
       }
       throw;
     }