+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_options.hpp
@@ -12,6 +12,7 @@
 #include <azure/storage/blobs/blob_options.hpp>
 #include <azure/storage/common/access_conditions.hpp>
 
+#include "azure/storage/files/datalake/datalake_responses.hpp"
 #include "azure/storage/files/datalake/protocol/datalake_rest_client.hpp"
 
 namespace Azure { namespace Storage { namespace Files { namespace DataLake {
@@ -209,9 +210,53 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
   };
 
   /**
-   * @brief Optional parameters for PathClient::AppendData
+   * @brief Optional parameters for FileSystemClient::GetAccessPolicy.
    */
-  struct AppendDataLakeFileDataOptions
+  struct GetDataLakeFileSystemAccessPolicyOptions
+  {
+    /**
+     * @brief Context for cancelling long running operations.
+     */
+    Azure::Core::Context Context;
+
+    /**
+     * @brief Optional conditions that must be met to perform this operation.
+     */
+    LeaseAccessConditions AccessConditions;
+  };
+
+  /**
+   * @brief Optional parameters for FileSystemClient::SetAccessPolicy.
+   */
+  struct SetDataLakeFileSystemAccessPolicyOptions
+  {
+    /**
+     * @brief Context for cancelling long running operations.
+     */
+    Azure::Core::Context Context;
+
+    /**
+     * @brief Specifies whether data in the file system may be accessed publicly and the level
+     * of access.
+     */
+    Models::PublicAccessType AccessType = Models::PublicAccessType::None;
+
+    /**
+     * @brief Stored access policies that you can use to provide fine grained control over
+     * file system permissions.
+     */
+    std::vector<Models::DataLakeSignedIdentifier> SignedIdentifiers;
+
+    /**
+     * @brief Optional conditions that must be met to perform this operation.
+     */
+    FileSystemAccessConditions AccessConditions;
+  };
+
+  /**
+   * @brief Optional parameters for PathClient::Append
+   */
+  struct AppendDataLakeFileOptions
   {
     /**
      * @brief Context for cancelling long running operations.
@@ -230,9 +275,9 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
   };
 
   /**
-   * @brief Optional parameters for PathClient::FlushData
+   * @brief Optional parameters for PathClient::Flush
    */
-  struct FlushDataLakeFileDataOptions
+  struct FlushDataLakeFileOptions
   {
     /**
      * @brief Context for cancelling long running operations.
@@ -283,9 +328,9 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
   };
 
   /**
-   * @brief Optional parameters for PathClient::SetAccessControl
+   * @brief Optional parameters for PathClient::SetAccessControlList
    */
-  struct SetDataLakePathAccessControlOptions
+  struct SetDataLakePathAccessControlListOptions
   {
     /**
      * @brief Context for cancelling long running operations.
@@ -303,13 +348,30 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     Azure::Core::Nullable<std::string> Group;
 
     /**
-     * @brief only valid if Hierarchical Namespace is enabled for the account. Sets POSIX
-     *        access permissions for the file owner, the file owning group, and others.
-     *        Each class may be granted read, write, or execute permission.
-     *        The sticky bit is also supported.  Both symbolic (rwxrw-rw-) and 4-digit octal
-     *        notation (e.g. 0766) are supported.
+     * @brief Specify the access condition for the path.
      */
-    Azure::Core::Nullable<std::string> Permissions;
+    PathAccessConditions AccessConditions;
+  };
+
+  /**
+   * @brief Optional parameters for PathClient::SetPermissions
+   */
+  struct SetDataLakePathPermissionsOptions
+  {
+    /**
+     * @brief Context for cancelling long running operations.
+     */
+    Azure::Core::Context Context;
+
+    /**
+     * @brief The owner of the path or directory.
+     */
+    Azure::Core::Nullable<std::string> Owner;
+
+    /**
+     * @brief The owning group of the path or directory.
+     */
+    Azure::Core::Nullable<std::string> Group;
 
     /**
      * @brief Specify the access condition for the path.