+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_responses.hpp
@@ -22,6 +22,17 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake { nam
   // FileSystemClient models:
 
   using ListPathsSinglePageResult = FileSystemListPathsResult;
+  using DataLakeSignedIdentifier = Blobs::Models::BlobSignedIdentifier;
+
+  struct GetDataLakeFileSystemAccessPolicyResult
+  {
+    PublicAccessType AccessType = PublicAccessType::None;
+    std::string ETag;
+    Azure::Core::DateTime LastModified;
+    std::vector<DataLakeSignedIdentifier> SignedIdentifiers;
+  }; // struct GetDataLakeFileSystemAccessPolicyResult
+
+  using SetDataLakeFileSystemAccessPolicyResult = Blobs::Models::SetBlobContainerAccessPolicyResult;
 
   struct GetDataLakeFileSystemPropertiesResult
   {
@@ -122,6 +133,9 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake { nam
   {
     std::string ETag;
     Core::DateTime LastModified;
+    std::string Owner;
+    std::string Group;
+    std::string Permissions;
     std::vector<Acl> Acls;
   };
 
@@ -145,13 +159,14 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake { nam
     Azure::Core::Nullable<int64_t> ContentLength;
   };
 
-  using SetDataLakePathAccessControlResult = PathSetAccessControlResult;
+  using SetDataLakePathAccessControlListResult = PathSetAccessControlResult;
+  using SetDataLakePathPermissionsResult = PathSetAccessControlResult;
 
   // FileClient models:
 
   using UploadDataLakeFileFromResult = Blobs::Models::UploadBlockBlobResult;
-  using AppendDataLakeFileDataResult = PathAppendDataResult;
-  using FlushDataLakeFileDataResult = PathFlushDataResult;
+  using AppendDataLakeFileResult = PathAppendDataResult;
+  using FlushDataLakeFileResult = PathFlushDataResult;
   using ScheduleDataLakeFileDeletionResult = Blobs::Models::SetBlobExpiryResult;
 
   struct ReadDataLakeFileResult