+++ sdk/core/azure-core/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Make `ToLower` and `LocaleInvariantCaseInsensitiveEqual` internal by moving them from `Azure::Core::Strings` to `Azure::Core::Internal::Strings`.
 
+### Bug Fixes
+
+- Fixed the parsing of the last chunk of a chunked response when using the curl transport adapter.
+
 ## 1.0.0-beta.4 (2021-01-13)
 
 ### New Features