+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release History
 
+## 12.0.0-beta.7 (Unreleased)
+
+### New Features
+
+- Added `Owner`, `Permissions`, and `Group` to `GetDataLakePathAccessControlResult`.
+- Added support for `GetAccessPolicy` and `SetAccessPolicy` in `DataLakeFileSystemClient`.
+
+### Breaking Changes
+
+- Removed `GetDfsUri` in all clients since they are currently implementation details.
+- Removed `Data` suffix for `FlushData` and `AppendData` and modified all related structs to align the change.
+- `DataLakePathClient` can no longer set permissions with `SetAccessControl`, instead, a new API `SetPermissions` is created for such functionality. Renamed the original API to `SetAccessControlList` to be more precise.
+- Renamed `GetUri` to `GetUrl`.
+
 ## 12.0.0-beta.6 (2020-01-14)
 
 ### New Features